+++ RealmSwift-swift1.2/List.swift
@@ -304,6 +304,22 @@ public final class List<T: Object>: ListBase {
     }
 
     /**
+    Moves the object at the given source index to the given destination index.
+
+    :warning: This method can only be called during a write transaction.
+    :warning: Throws an exception when called with an index smaller than zero or greater than
+              or equal to the number of objects in the list.
+
+    :param: from  The index of the object to be moved.
+    :param: to    The index to which the object at `from` should be moved.
+    */
+    public func move(#from: Int, to: Int) {
+        throwForNegativeIndex(from)
+        throwForNegativeIndex(to)
+        _rlmArray.moveObjectAtIndex(UInt(from), toIndex: UInt(to))
+    }
+
+    /**
     Exchanges the objects in the list at given indexes.
 
     :warning: Throws an exception when either index exceeds the bounds of the list.